update: use ts and bun.
