+++ .gitignore
@@ -134,3 +134,13 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+
+# macOS
+.DS_Store
+
+# IDE
+.idea/
+
+# Test artifacts
+test-screenshots/
+comparison/
