+++ eslint.config.js
@@ -1,9 +1,29 @@
 import globals from "globals";
 import pluginJs from "@eslint/js";
-
+import tseslint from "typescript-eslint";
 
 /** @type {import('eslint').Linter.Config[]} */
 export default [
-  {languageOptions: { globals: globals.node }},
+  {
+    files: ["**/*.{js,mjs,cjs,ts}"],
+    languageOptions: {
+      globals: {
+        ...globals.node,
+        Bun: "readonly",
+      }
+    }
+  },
   pluginJs.configs.recommended,
+  ...tseslint.configs.recommended,
+  {
+    rules: {
+      "@typescript-eslint/no-unused-vars": [
+        "error",
+        {
+          argsIgnorePattern: "^_",
+          varsIgnorePattern: "^_",
+        }
+      ]
+    }
+  }
 ];
