+++ package.json
@@ -2,10 +2,13 @@
 	"name": "appwrite-browser",
 	"version": "0.1.0",
 	"description": "",
-	"main": "src/index.js",
+	"main": "src/server.ts",
 	"type": "module",
 	"scripts": {
-		"start": "node src/index.js",
+		"start": "bun run src/server.ts",
+		"dev": "bun --watch src/server.ts",
+		"build": "tsc",
+		"type-check": "tsc --noEmit",
 		"format": "biome check --write ./src",
 		"lint": "biome check ./src",
 		"check": "eslint ./src",
@@ -16,7 +19,6 @@
 	"license": "ISC",
 	"packageManager": "pnpm@10.20.0",
 	"dependencies": {
-		"h3": "^1.13.0",
 		"lighthouse": "^12.2.1",
 		"playwright-core": "^1.52.0",
 		"playwright-lighthouse": "^4.0.0",
@@ -25,7 +27,10 @@
 	"devDependencies": {
 		"@biomejs/biome": "^1.9.4",
 		"@eslint/js": "^9.20.0",
+		"@types/bun": "^1.3.2",
 		"eslint": "^9.20.1",
-		"globals": "^15.15.0"
+		"globals": "^15.15.0",
+		"typescript": "^5.9.3",
+		"typescript-eslint": "^8.46.3"
 	}
 }
