+++ src/config/index.ts
@@ -1,0 +1,4 @@
+export * from "./browser.js";
+export * from "./lighthouse.js";
+
+export const port = Number(process.env.PORT) || 3000;
