+++ src/routes/index.ts
@@ -1,0 +1,4 @@
+export * from "./test.js";
+export * from "./health.js";
+export * from "./reports.js";
+export * from "./screenshots.js";
