+++ src/schemas/index.ts
@@ -1,0 +1,2 @@
+export * from "./screenshot.schema.js";
+export * from "./lighthouse.schema.js";
